# laravel/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Swagger with Laravel",
        "version": "1.0.0"
    },
    "paths": {
        "/api/artisans": {
            "get": {
                "tags": [
                    "Artisan"
                ],
                "summary": "Get all Artisans",
                "description": "Display a listing of the resource.",
                "operationId": "4997942c6261ddbf23a4879bcf598e0e",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Artisan"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Artisan"
                ],
                "summary": "Create a new Artisan",
                "description": "Store a newly created resource in storage.",
                "operationId": "7a589e653b95179b9169cee7359a2ff9",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Artisan"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Artisan created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Artisan"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/artisans/{id}": {
            "get": {
                "tags": [
                    "Artisan"
                ],
                "summary": "Get a specific Artisan by ID",
                "description": "Display the specified resource.",
                "operationId": "0c08e1b4e8f3ba57f7778bcba63be4cf",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Artisan ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Artisan"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Artisan not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Artisan"
                ],
                "summary": "Update a specific Artisan by ID",
                "description": "Update the specified resource in storage.",
                "operationId": "a68d608734b30c4384a89627c88e0cb5",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Artisan ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Artisan"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Artisan updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Artisan"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Artisan not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Artisan"
                ],
                "summary": "Delete a specific Artisan by ID",
                "description": "Remove the specified resource from storage.",
                "operationId": "98ae49f03469266ed3a8f1df86414453",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Artisan ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Artisan deleted successfully"
                    },
                    "404": {
                        "description": "Artisan not found"
                    }
                }
            }
        },
...
        "/api/consumers": {
            "get": {
                "tags": [
                    "Consumer"
                ],
                "summary": "Get all Consumers",
                "description": "Display a listing of the resource.",
                "operationId": "39606af6450e02dac37b094b7e141744",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Consumer"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Consumer"
                ],
                "summary": "Create a new Consumer",
                "description": "Store a newly created resource in storage.",
                "operationId": "e89bd3292e20158fb85d8b85ed25cb90",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Consumer"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Consumer created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Consumer"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/consumers/{id}": {
            "get": {
                "tags": [
                    "Consumer"
                ],
                "summary": "Get a specific Consumer by ID",
                "description": "Display the specified resource.",
                "operationId": "2d4557cac212e72d79c67bf715ba7b3d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Consumer ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Consumer"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Consumer not found"
                    }
                }
            }
        },
        "/api/consumers/{id}": {
            "put": {
                "tags": [
                    "Consumer"
                ],
                "summary": "Update a specific Consumer by ID",
                "description": "Update the specified resource in storage.",
                "operationId": "034ae4e87f1bc77a755cd0bc37db0f0b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Consumer ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Consumer"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Consumer updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Consumer"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Consumer not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Consumer"
                ],
                "summary": "Delete a specific Consumer by ID",
                "description": "Remove the specified resource from storage.",
                "operationId": "320b22d41bfc9d9102ad5f7fda3af395",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Consumer ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Consumer deleted successfully"
                    },
                    "404": {
                        "description": "Consumer not found"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "Artisan": {
                "properties": {
                    "id": {
                        "description": "Artisan ID",
                        "type": "integer",
                        "format": "int64"
                    },
                    "created_at": {
                        "description": "Artisan creation timestamp",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "description": "Artisan update timestamp",
                        "type": "string",
                        "format": "date-time"
                    },
                    "user": {
                        "$ref": "#/components/schemas/User"
                    }
                },
                "type": "object"
            },
            "Consumer": {
                "properties": {
                    "id": {
                        "description": "Consumer ID",
                        "type": "integer",
                        "format": "int64"
                    },
                    "created_at": {
                        "description": "Consumer creation timestamp",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "description": "Consumer update timestamp",
                        "type": "string",
                        "format": "date-time"
                    },
                    "user": {
                        "$ref": "#/components/schemas/User"
                    },
                    "order": {
                        "$ref": "#/components/schemas/Order"
                    }
                },
                "type": "object"
            },
            "DeliveryPersonnel": {
                "properties": {
                    "id": {
                        "description": "DeliveryPersonnel ID",
                        "type": "integer",
                        "format": "int64"
                    },
                    "created_at": {
                        "description": "DeliveryPersonnel creation timestamp",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "description": "DeliveryPersonnel update timestamp",
                        "type": "string",
                        "format": "date-time"
                    },
                    "user": {
                        "$ref": "#/components/schemas/User"
                    }
                },
                "type": "object"
            },
            "Order": {
                "properties": {
                    "id": {
                        "description": "Order ID",
                        "type": "integer",
                        "format": "int64"
                    },
                    "orderDate": {
                        "description": "Order date",
                        "type": "string",
                        "format": "date-time"
                    },
                    "orderStatus": {
                        "description": "Order status",
                        "type": "string"
                    },
                    "deliveryAddress": {
                        "description": "Delivery address",
                        "type": "string"
                    },
                    "consumer_id": {
                        "description": "ID of the associated consumer",
                        "type": "integer",
                        "format": "int64"
                    },
                    "created_at": {
                        "description": "Order creation timestamp",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "description": "Order update timestamp",
                        "type": "string",
                        "format": "date-time"
                    },
                    "deliveryPersonnel": {
                        "$ref": "#/components/schemas/DeliveryPersonnel"
                    },
                    "consumer": {
                        "$ref": "#/components/schemas/Consumer"
                    },
                    "orderProducts": {
                        "description": "Related OrderProducts",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/OrderProduct"
                        }
                    }
                },
                "type": "object"
            },
            "OrderProduct": {
                "properties": {
                    "id": {
                        "description": "OrderProduct ID",
                        "type": "integer",
                        "format": "int64"
                    },
                    "quantity": {
                        "description": "OrderProduct quantity",
                        "type": "integer"
                    },
                    "order": {
                        "$ref": "#/components/schemas/Order"
                    },
                    "product": {
                        "$ref": "#/components/schemas/Product"
                    },
                    "created_at": {
                        "description": "OrderProduct creation timestamp",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "description": "OrderProduct update timestamp",
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "Product": {
                "properties": {
                    "id": {
                        "description": "Product ID",
                        "type": "integer",
                        "format": "int64"
                    },
                    "name": {
                        "description": "Product name",
                        "type": "string"
                    },
                    "description": {
                        "description": "Product description",
                        "type": "string"
                    },
                    "pricePerPiece": {
                        "description": "Product price per piece",
                        "type": "number",
                        "format": "float"
                    },
                    "minOrder": {
                        "description": "Minimum order quantity",
                        "type": "integer"
                    },
                    "type": {
                        "description": "Product type",
                        "type": "string",
                        "enum": [
                            "Sugar",
                            "Salt"
                        ]
                    },
                    "childType": {
                        "description": "Product child type",
                        "type": "string"
                    },
                    "images": {
                        "description": "Product images",
                        "type": "json"
                    },
                    "ratings": {
                        "description": "Related Ratings",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Rating"
                        }
                    },
                    "artisan": {
                        "$ref": "#/components/schemas/Artisan"
                    },
                    "orderProducts": {
                        "description": "Related OrderProducts",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/OrderProduct"
                        }
                    },
                    "created_at": {
                        "description": "Product creation timestamp",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "description": "Product update timestamp",
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "Rating": {
                "properties": {
                    "id": {
                        "description": "Rating ID",
                        "type": "integer",
                        "format": "int64"
                    },
                    "rating": {
                        "description": "Numeric rating value",
                        "type": "integer"
                    },
                    "ratingType": {
                        "description": "Type of rating (DeliveryPersonnel, Product, Artisan)",
                        "type": "string",
                        "enum": [
                            "DeliveryPersonnel",
                            "Product",
                            "Artisan"
                        ]
                    },
                    "created_at": {
                        "description": "Rating creation timestamp",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "description": "Rating update timestamp",
                        "type": "string",
                        "format": "date-time"
                    },
                    "delivery_personnel": {
                        "$ref": "#/components/schemas/DeliveryPersonnel"
                    },
                    "product": {
                        "$ref": "#/components/schemas/Product"
                    },
                    "artisan": {
                        "$ref": "#/components/schemas/Artisan"
                    }
                },
                "type": "object"
            },
            "User": {
                "properties": {
                    "id": {
                        "description": "User ID",
                        "type": "integer",
                        "format": "int64"
                    },
                    "email": {
                        "description": "User email",
                        "type": "string"
                    },
                    "first_name": {
                        "description": "User first name",
                        "type": "string"
                    },
                    "last_name": {
                        "description": "User last name",
                        "type": "string"
                    },
                    "description": {
                        "description": "User description",
                        "type": "string"
                    },
                    "adresse": {
                        "description": "User address",
                        "type": "string"
                    },
                    "phone_number": {
                        "description": "User phone number",
                        "type": "string"
                    },
                    "user_type": {
                        "description": "User type",
                        "type": "string",
                        "enum": [
                            "Artisan",
                            "Consumer",
                            "DeliveryPersonnel"
                        ]
                    },
                    "created_at": {
                        "description": "User creation timestamp",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "description": "User update timestamp",
                        "type": "string",
                        "format": "date-time"
                    },
                    "deliveryPersonnel": {
                        "$ref": "#/components/schemas/DeliveryPersonnel"
                    },
                    "artisan": {
                        "$ref": "#/components/schemas/Artisan"
                    },
                    "consumer": {
                        "$ref": "#/components/schemas/Consumer"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}